Added basic search capacity, and add to fix lots of things for that.
- Linking between quote and authors, not only between source and authors.
E.g., in a collective work, a quote might be from a particular author.

- That meant adjusting the parser so that author could be specified for quotes

- The 'shortcut', business logic of db insertion, had to be changed too.

- Also various queries and db methods.

This means the serialization framework is less and less readable though.
And we have issues for multi-author work.
And we should get more data for a source.
This should be easy to fix with some work, but at the cost of more
bloated serialization framework methods.

But first : let's fix the tests.